requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.3
supabase>=2.0.0
pypdf>=3.17.1
tiktoken>=0.5.2
openai>=1.3.7
tqdm>=4.66.1
requests>=2.31.0
streamlit>=1.29.0
python-dotenv>=1.0.0